lockscreen: show device name
It is useful to see the name of the device before you unlock it. For
example, if you have multiple BitBoxes, it is useful to know which one
is inserted before entering long passwords or passphrases.

The device name is not a secret, and can already be queried in the API
at any time, also before unlock.

Users that dislike this can change the device name in the settings to
something generic.